add social media

+++ app/Modules/User/Config/Routes.php
@@ -8,11 +8,11 @@ $routes->group('admin', ['namespace' => 'App\Modules\User\Controllers', 'filter'
 
     /*** Route for User ***/
     $subroutes->get('user', 'User::index', ['filter' => 'authAdmin']);
-    $subroutes->get('user/add', 'User::create');
-    $subroutes->post('user', 'User::store');
-    $subroutes->get('user/edit/(:any)', 'User::edit/$1');
-    $subroutes->put('user/(:any)', 'User::update/$1');
-    $subroutes->delete('user/(:any)', 'User::destroy/$1');
+    $subroutes->get('user/add', 'User::create', ['filter' => 'authAdmin']);
+    $subroutes->post('user', 'User::store', ['filter' => 'authAdmin']);
+    $subroutes->get('user/edit/(:any)', 'User::edit/$1', ['filter' => 'authAdmin']);
+    $subroutes->put('user/(:any)', 'User::update/$1', ['filter' => 'authAdmin']);
+    $subroutes->delete('user/(:any)', 'User::destroy/$1', ['filter' => 'authAdmin']);
 });
 
 $routes->group('admin/api', ['namespace' => 'App\Modules\User\Controllers', 'filter' => 'authGuard'], function ($subroutes) {
